+++ database/migrations/2018_10_24_193205_stuff2.php
@@ -15,10 +15,10 @@ class Stuff2 extends Migration
     {
         //
         Schema::create('comments', function(Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->morphs('commentable');
             $table->text('content');
-            $table->unsignedInteger('likes');
+            $table->unsignedBigInteger('likes');
             $table->timestamps();
         });
     }
